✨ Authentification refacto with fetchJWT util file
